requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==24.1.0
    # via
    #   -r requirements.in
    #   aiogram
aiogram==3.22.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via
    #   -r requirements.in
    #   aiogram
    #   aiohttp-socks
aiohttp-socks==0.10.1
    # via -r requirements.in
aioredis==2.0.1
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
alembic==1.16.5
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
apscheduler==3.11.0
    # via -r requirements.in
async-timeout==5.0.1
    # via aioredis
attrs==25.3.0
    # via aiohttp
bcrypt==4.3.0
    # via
    #   -r requirements.in
    #   passlib
certifi==2025.8.3
    # via
    #   aiogram
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   typer
    #   uvicorn
cryptography==46.0.1
    # via
    #   -r requirements.in
    #   python-jose
ecdsa==0.19.1
    # via python-jose
fastapi==0.116.2
    # via
    #   -r requirements.in
    #   fastapi-pagination
fastapi-pagination==0.14.1
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
grpcio==1.62.0
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.62.0
    # via -r requirements.in
grpclib==0.4.8
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
h2==4.3.0
    # via grpclib
hpack==4.1.0
    # via h2
httptools==0.6.4
    # via -r requirements.in
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jalali-core==1.0.0
    # via jdatetime
jdatetime==5.2.0
    # via -r requirements.in
jinja2==3.1.6
    # via -r requirements.in
magic-filter==1.0.12
    # via aiogram
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
multidict==6.6.4
    # via
    #   aiohttp
    #   grpclib
    #   yarl
packaging==25.0
    # via pytest
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==4.25.2
    # via
    #   -r requirements.in
    #   grpcio-tools
psycopg==3.2.10
    # via -r requirements.in
psycopg-binary==3.2.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   aiogram
    #   fastapi
    #   fastapi-pagination
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via -r requirements.in
pymysql==1.1.2
    # via -r requirements.in
pynacl==1.6.0
    # via -r requirements.in
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==1.2.0
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
python-socks[asyncio]==2.7.2
    # via aiohttp-socks
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   v2share
requests==2.32.5
    # via -r requirements.in
rich==14.1.0
    # via
    #   -r requirements.in
    #   typer
rsa==4.9.1
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
starlette==0.48.0
    # via
    #   -r requirements.in
    #   fastapi
typer==0.17.4
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   aiogram
    #   aioredis
    #   aiosignal
    #   alembic
    #   anyio
    #   fastapi
    #   fastapi-pagination
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzlocal==5.3.1
    # via apscheduler
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
v2share==0.1.0b31
    # via -r requirements.in
websockets==15.0.1
    # via -r requirements.in
xxhash==3.5.0
    # via -r requirements.in
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
aiofiles
aiogram
aiohttp
aiohttp-socks
aioredis
alembic
apscheduler
bcrypt
cryptography
fastapi
fastapi-pagination
grpcio
grpcio-tools
grpclib
httptools
jdatetime
jinja2
passlib
protobuf
psycopg
psycopg-binary
pydantic
pyjwt
pymysql
pynacl
pytest
pytest-asyncio
python-decouple
python-dotenv
python-jose
python-multipart
python-socks
pyyaml
requests
rich
sqlalchemy
starlette
typer
uvicorn
v2share
websockets
xxhash
